Rename all camelCase files to kebab-case
