+++ app.js
@@ -40,6 +40,4 @@ app.use(function(err, req, res, next) {
   res.render('error');
 });
 
-app.listen(5000);
-
 module.exports = app;
